perf(ssh): gf_balance_s16 in int32 not int64 (X25519 1.55x, ed25519 1.44x)
The S3 vector multiply balances each operand into signed-16-bit before the ACCX
convolution. That balance ran in int64, so its 48-step carry propagation (3 passes
x 16 limbs, per operand) was emulated 64-bit math on the 32-bit Xtensa - and it
dominated the field op. The limbs stay ~+-2^18 and carries ~+-2, well inside int32,
so int32 is byte-exact and skips the emulation.

Device-measured: X25519 115.9 -> 97.5 ms (1.55x vs scalar), ed25519_sign
463.3 -> 380.3 ms (1.44x); handshake crypto ~0.85 -> ~0.58 s. Byte-exact 3000/3000
on device, interop 6/6 vs OpenSSH, native gf_mul_s16 model matches scalar over 20k
random operands.

This also settled the load question: an in-register sliding window that removes
every per-output ee.ld.128.usar is byte-exact but only ~1.4% faster (7955 vs 8067
cyc), i.e. the window loads were never the bottleneck - the int64 balance was. Not
shipping the window rewrite (complexity without payoff); balance was the real lever.

+++ src/network_drivers/presentation/ssh/crypto/ssh_curve25519.cpp
@@ -66,15 +66,17 @@ void ssh_gf_sub(ssh_gf out, const ssh_gf a, const ssh_gf b)
 // overflow wraps *38 into limb 0 since 2^256 == 38; three passes settle the wrap).
 static void gf_balance_s16(int16_t o[16], const ssh_gf a)
 {
-    int64_t c[16];
+    // int32 throughout: limbs stay ~+-2^18 and carries ~+-2, so no value exceeds int32 - which avoids the
+    // emulated 64-bit carry-propagation math (48 steps per operand) that dominated the field multiply.
+    int32_t c[16];
     for (int i = 0; i < 16; i++)
-        c[i] = a[i];
+        c[i] = (int32_t)a[i];
     for (int pass = 0; pass < 3; pass++)
     {
-        int64_t carry = 0;
+        int32_t carry = 0;
         for (int i = 0; i < 16; i++)
         {
-            int64_t v = c[i] + carry;
+            int32_t v = c[i] + carry;
             carry = (v + 0x8000) >> 16;
             c[i] = v - (carry << 16);
         }
